Use local meta solutions
After type checking a definition we now push the meta solutions as deep
into the resulting term as possible (while still covering all
occurrences), instead of keeping them at the top level.

The benefit to this is that the meta solution often ends up in a context
where many of the arguments to the meta variables can be inlined.

It also means that the query system doesn't need to care about metas,
because they're in the terms.