Make case expression elaboration work
This required some changes to how we handle variables in the context,
because dependent pattern matching elaboration requires changes in the
middle of the context, which is something that is annoying to handle
with De Bruijn indices at the syntax level.

So the fix is to decouple the De Bruijn indices from their position in
the context. We always introduce new variables with index 0, but index 0
might map to a variable that's not at index 0. This is what the
Index.Map module does for us.