Keep track of multiple source spans per binding
Some names might have multiple source spans, because they stem from
multiple clauses in a definition that has been desugared. We would
previously just use the first source span for such names, meaning that
jump to definition was a bit buggy in the presence of pattern matching.

With these changes we keep track of potentially many source spans, and
when doing jump to definition we try to pick the right one by using the
closest one before the cursor position.