Add the beginnings of an LLVM backend
Compile the closure converted syntax first to a home-made assembly
language (Assembly) similar to LLVM (but with less type casting), then
to a CPS'ed form of the same assembly (CPSAssembly) which makes the
stack operations explicit.

Compile CPSAssembly to LLVM using only tail calls and the GHC calling
convention. This means that it doesn't use the C stack, but handles the
stack itself (even though LLVM normally handles that). This is important
since I'd like to have full control over its layout to implement stack
scanning for GC.

This is missing a bunch of features (data types, heap allocations, GC,
compilation of the generated code, etc).