Limit gluing to lets that are bound outside of the current evaluation
Keep track of an index/level in the environment to the left of which
gluing is allowed.

The idea is that let bindings that you encounter during evaluation can't
be glued because that'd mean that the variable escapes. But lets that
are bound during elaboration can be glued because then you're working
relative to a context where those variables are bound.